+++ Assets/Scripts/ScriptableObjectScripts/EnemyData.cs
@@ -4,6 +4,6 @@ public class EnemyData : CharData
     public float ContactDamage;
     public float TimeBeforeNoticePlayer;
     public bool _stunned;
-    public float PuntuationXHit;
-    public float PuntuationXDeath;
+    public int PuntuationXHit;
+    public int PuntuationXDeath;
 }
